Update nginx and fix redirection issue
Nginx 301 redirects url without trailing slash to url with trailing
slash, but on wrong portm, because it listens on 8443 while the exposed
port in k8s is 443.

New bitnami docker image adds new environment variable
NGINX_ENABLE_ABSOLUTE_REDIRECT and NGINX_ENABLE_PORT_IN_REDIRECT, both
defaulting to false. So the default is to use relative redirect, which
doesn't include port in it.