base: rework node representation in code
It is now based on NodePatch, such that we can set additional labels and
annotations on nodes.

It is not possible to directly import Node as resource in pulumi,
because there are labels and annotations added automatically by node
feature discovery, so there is no way to set all of them ahead of time
in code.

It's also not worth it to always get Node afresh every time evaulating
the code, which is very slow.